cli is added code refactoring
