+++ Udemy_Advance_Python/App-2-Flatmates-Bill/resources/design.txt
